requirements.txt
anthropic==0.76.0
numpy==2.3.2
pandas==2.3.2
py_vollib==1.0.1
PyYAML==6.0.3
rich==14.3.1
scipy==1.17.1
textual==8.1.1
yfinance==1.1.0
